task-service/requirements.txt
flask
psycopg2-binary
pyjwt
flask_cors